# app/server/src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "WebRTC Calling Application API",
    "version": "2.2.4",
    "description": "Audio calling application with WebRTC signaling server. Provides authentication, real-time communication via WebSocket, and call management.",
    "contact": {
      "name": "API Support",
      "url": "https://github.com/your-repo/calling-app"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://api.yourapp.com",
      "description": "Production server"
    }
  ],
  "tags": [
    {
      "name": "Health",
      "description": "Health check and system status endpoints"
    },
    {
      "name": "Authentication",
      "description": "User authentication and token management"
    },
    {
      "name": "WebSocket",
      "description": "Real-time WebSocket communication for calls"
    }
  ],
  "paths": {
    "/": {
      "get": {
        "tags": ["Health"],
        "summary": "Health check",
        "description": "Returns server status, version, and enabled features",
        "responses": {
          "200": {
            "description": "Server is healthy",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                },
                "example": {
                  "service": "Audio Calling Web App - Signaling Server",
                  "status": "running",
                  "timestamp": "2024-12-18T12:00:00.000Z",
                  "version": "2.2.4",
                  "environment": "development",
                  "features": [
                    "authentication",
                    "rate-limiting",
                    "input-validation",
                    "connection-cleanup",
                    "helmet-security-headers",
                    "strict-cors",
                    "env-validation",
                    "winston-logging"
                  ],
                  "security": {
                    "https": false,
                    "guestAccess": true
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/health/redis": {
      "get": {
        "tags": ["Health"],
        "summary": "Redis health check",
        "description": "Check Redis connection status and get Redis info",
        "responses": {
          "200": {
            "description": "Redis is healthy or disabled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RedisHealthResponse"
                }
              }
            }
          },
          "503": {
            "description": "Redis is unhealthy",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/guest": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Guest authentication",
        "description": "Create a guest user session and receive JWT token. Rate limited to 5 requests per 15 minutes.",
        "responses": {
          "200": {
            "description": "Successfully authenticated as guest",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                },
                "example": {
                  "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                  "extension": "123-456",
                  "expiresIn": "24h"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "User login",
        "description": "Authenticate with username and password. Rate limited to 5 requests per 15 minutes.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              },
              "example": {
                "username": "user@example.com",
                "password": "securePassword123"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many login attempts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "tags": ["Authentication"],
        "summary": "User registration",
        "description": "Register a new user account. Rate limited to 5 requests per 15 minutes.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              },
              "example": {
                "username": "newuser@example.com",
                "password": "securePassword123",
                "confirmPassword": "securePassword123"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many registration attempts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HealthResponse": {
        "type": "object",
        "properties": {
          "service": {
            "type": "string",
            "example": "Audio Calling Web App - Signaling Server"
          },
          "status": {
            "type": "string",
            "enum": ["running", "degraded", "down"],
            "example": "running"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "version": {
            "type": "string",
            "example": "2.2.4"
          },
          "environment": {
            "type": "string",
            "enum": ["development", "production", "test"],
            "example": "development"
          },
          "features": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "security": {
            "type": "object",
            "properties": {
              "https": {
                "type": "boolean"
              },
              "guestAccess": {
                "type": "boolean"
              }
            }
          }
        }
      },
      "RedisHealthResponse": {
        "type": "object",
        "properties": {
          "enabled": {
            "type": "boolean"
          },
          "status": {
            "type": "string",
            "enum": ["healthy", "unhealthy", "disabled"]
          },
          "connected": {
            "type": "boolean"
          },
          "usedMemory": {
            "type": "string"
          },
          "connectedClients": {
            "type": "integer"
          }
        }
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "JWT token for authentication"
          },
          "extension": {
            "type": "string",
            "pattern": "^\\d{3}-\\d{3}$",
            "example": "123-456",
            "description": "User's assigned extension number"
          },
          "expiresIn": {
            "type": "string",
            "example": "24h"
          }
        },
        "required": ["token", "extension"]
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 8
          }
        },
        "required": ["username", "password"]
      },
      "RegisterRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 8
          },
          "confirmPassword": {
            "type": "string",
            "minLength": 8
          }
        },
        "required": ["username", "password", "confirmPassword"]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string"
              },
              "status": {
                "type": "integer"
              },
              "stack": {
                "type": "string",
                "description": "Only included in development mode"
              }
            }
          }
        }
      },
      "WebSocketEvents": {
        "type": "object",
        "description": "WebSocket events for real-time communication",
        "properties": {
          "call-user": {
            "type": "object",
            "description": "Initiate a call to another user",
            "properties": {
              "targetExtension": {
                "type": "string",
                "pattern": "^\\d{3}-\\d{3}$"
              },
              "offer": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": ["offer"]
                  },
                  "sdp": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "answer-call": {
            "type": "object",
            "description": "Answer an incoming call",
            "properties": {
              "targetExtension": {
                "type": "string"
              },
              "answer": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": ["answer"]
                  },
                  "sdp": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "ice-candidate": {
            "type": "object",
            "description": "Exchange ICE candidates for WebRTC connection",
            "properties": {
              "targetExtension": {
                "type": "string"
              },
              "candidate": {
                "type": "object"
              }
            }
          },
          "end-call": {
            "type": "object",
            "description": "End an active call",
            "properties": {
              "targetExtension": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from authentication endpoints"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}